Implemented new setSubTaskParent api call to add or remove a parent
Described by the api update as:
Subtask re-homing. You can set the parent field on a Task to change which Task
it is a subtask of.

Use 'null' to remove a parent from a subtask with setSubTaskParent

Allow removal of a parent from a subtask to make it a normal task by
using 'null' for the parent value. This needs to be quoted, otherwise
the parent field will be removed from the body and the API will return
a 400 bad request